Update ForApproxValueOf e .c

+++ ForApproxValueOf e .c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 void main()
 {
-    float f=1,e=0,a=1;
+    long double f=1,e=0,a=1;
     int i,j,n;
     printf("Estimate value of e till : ");
     scanf("%d",&n);
@@ -14,5 +14,5 @@ void main()
             e = e + 1.0/f;
         }
     }
-    printf("e = %f",e);
+    printf("e = %Lf",e);
 }
